requirements.txt
fastapi
uvicorn
pypdf
google-generativeai
streamlit
pytest
requests
